+++ source/lambda/services/reporter/__tests__/limit-report.spec.ts
@@ -7,6 +7,7 @@ import { UnsupportedEventException } from "solutions-utils";
 const deleteMessageMock = jest.fn();
 const receiveMessagesMock = jest.fn();
 const putItemMock = jest.fn();
+const destroyMock = jest.fn();
 
 jest.mock("solutions-utils", () => {
   const originalModule = jest.requireActual("solutions-utils");
@@ -17,6 +18,7 @@ jest.mock("solutions-utils", () => {
       return {
         deleteMessage: deleteMessageMock,
         receiveMessages: receiveMessagesMock,
+        destroy: destroyMock,
       };
     },
     DynamoDBHelper: function () {
@@ -111,6 +113,7 @@ describe("limitreport", function () {
       expect(receiveMessagesMock).toHaveBeenCalledTimes(2);
       expect(deleteMessageMock).toHaveBeenCalledTimes(4);
       expect(putItemMock).toHaveBeenCalledTimes(4);
+      expect(destroyMock).toHaveBeenCalledTimes(2);
     });
 
     it("handler should throw error for unsupported trigger", async () => {
@@ -126,6 +129,7 @@ describe("limitreport", function () {
       expect(receiveMessagesMock).toHaveBeenCalledTimes(2);
       expect(deleteMessageMock).toHaveBeenCalledTimes(4);
       expect(putItemMock).toHaveBeenCalledTimes(4);
+      expect(destroyMock).toHaveBeenCalledTimes(2);
     });
 
     it("should log dynamo error when put fails", async () => {
@@ -134,6 +138,7 @@ describe("limitreport", function () {
       expect(receiveMessagesMock).toHaveBeenCalledTimes(2);
       expect(deleteMessageMock).toHaveBeenCalledTimes(3);
       expect(putItemMock).toHaveBeenCalledTimes(4);
+      expect(destroyMock).toHaveBeenCalledTimes(2);
     });
 
     it("should handle some message loops returning empty arrays", async () => {
@@ -146,6 +151,7 @@ describe("limitreport", function () {
       expect(receiveMessagesMock).toHaveBeenCalledTimes(2);
       expect(deleteMessageMock).toHaveBeenCalledTimes(2);
       expect(putItemMock).toHaveBeenCalledTimes(2);
+      expect(destroyMock).toHaveBeenCalledTimes(1);
     });
     it("should return an error if the message is empty", async () => {
       receiveMessagesMock.mockResolvedValue(emptyMessageBody);
@@ -155,6 +161,7 @@ describe("limitreport", function () {
       expect(receiveMessagesMock).toHaveBeenCalledTimes(2);
       expect(deleteMessageMock).toHaveBeenCalledTimes(0);
       expect(putItemMock).toHaveBeenCalledTimes(0);
+      expect(destroyMock).toHaveBeenCalledTimes(0);
     });
 
     it("should handle an error if the message fails to delete from the queue", async () => {
@@ -164,6 +171,7 @@ describe("limitreport", function () {
       expect(receiveMessagesMock).toHaveBeenCalledTimes(2);
       expect(deleteMessageMock).toHaveBeenCalledTimes(4);
       expect(putItemMock).toHaveBeenCalledTimes(4);
+      expect(destroyMock).toHaveBeenCalledTimes(2);
     });
 
     it("should handle missing environment variables", async () => {
@@ -179,6 +187,7 @@ describe("limitreport", function () {
       expect(receiveMessagesMock).toHaveBeenCalledTimes(0);
       expect(deleteMessageMock).toHaveBeenCalledTimes(0);
       expect(putItemMock).toHaveBeenCalledTimes(0);
+      expect(destroyMock).toHaveBeenCalledTimes(0);
     });
   });
 });