+++ source/resources/__tests__/test-context.ts
@@ -1,0 +1,19 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export default {
+  SOLUTION_VERSION: "test_version",
+  SOLUTION_NAME: "test_name",
+  SOLUTION_ID: "SO0005",
+  SOLUTION_BUCKET: "test_bucket",
+  SOLUTION_TEMPLATE_BUCKET: "test_bucket",
+  CUSTOM_USER_AGENT: "AwsSolution/SO0005/test_version",
+  SEND_METRICS: "Yes",
+  METRICS_ENDPOINT: "https://metrics.awssolutionsbuilder.com/generic",
+  LOG_LEVEL: "info",
+  APPLICATION_TYPE: "AWS-Solutions",
+  APP_REG_HUB_NO_OU_APPLICATION_NAME: "QM_Hub",
+  APP_REG_HUB_APPLICATION_NAME: "QM_Hub_Org",
+  APP_REG_TA_SPOKE_APPLICATION_NAME: "QM_TA",
+  APP_REG_SQ_SPOKE_APPLICATION_NAME: "QM_SQ",
+};