+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.2.5] - 2024-01-08
+
+### Changed
+- Made reporting of OK Messages optional
+- Added percentage marker on Service Quota notifications
+
+### Fixed
+- Corrected expiration dates of records in DynamoDb history table
+- Added manual resource cleanup after sqs message consumption
+
 ## [6.2.4] - 2023-11-09
 
 ### Changed